+++ lib/ood_core/job/adapters/systemd/launcher.rb
