+++ lib/ood_core/job/adapters/systemd/templates/email.erb.sh
