+++ lib/ood_core/job/adapters/systemd/templates/script_wrapper.erb.sh
